refactor: move channel requestListener to registrar
