+++ src/DependencyInjection/EMSCoreExtension.php
@@ -31,6 +31,7 @@ class EMSCoreExtension extends Extension implements PrependExtensionInterface
         $xmlLoader->load('repositories.xml');
         $yamlLoader->load('services.yml');
         $xmlLoader->load('controllers.xml');
+        $xmlLoader->load('services.xml');
 
         $container->setParameter('ems_core.from_email', $config['from_email']);
         $container->setParameter('ems_core.instance_id', $config['instance_id']);
