+++ src/Resources/config/services.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" ?>
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services
+        http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <parameters>
+        <parameter key="env(EMSCH_ENV)"/>
+    </parameters>
+
+    <services>
+        <defaults public="false" />
+
+        <!-- event listeners -->
+        <service id="ems.event_listener.request_listener" class="EMS\CoreBundle\EventListener\RequestListener">
+            <argument type="service" id="ems.service.channel.register"/>
+            <argument type="service" id="twig"/>
+            <argument type="service" id="doctrine"/>
+            <argument type="service" id="logger"/>
+            <argument type="service" id="router"/>
+            <tag name="kernel.event_listener" event="kernel.exception" method="onKernelException" />
+            <tag name="kernel.event_listener" event="kernel.request" method="onKernelRequest" priority="110" />
+        </service>
+
+        <!-- core service -->
+        <service id="ems.service.channel.register" class="EMS\CoreBundle\Service\Channel\ChannelRegistrar">
+            <argument type="service" id="ems.repository.channel"/>
+            <argument type="service" id="emsch.helper_environment"/>
+            <argument type="service" id="logger"/>
+            <argument type="service" id="ems.service.index"/>
+        </service>
+
+    </services>
+</container>
