+++ src/Resources/config/services.yml
@@ -27,15 +27,6 @@ services:
             - '@EMS\CoreBundle\Repository\UserRepository'
         tags:
             - { name: twig.runtime, priority: -2000 }
-    app.listener.request:
-        class: EMS\CoreBundle\EventListener\RequestListener
-        arguments: ['@twig', '@doctrine', '@logger', '@router', '@ems.repository.channel', '@emsch.helper_environment', '@ems.service.index']
-        tags:
-            #- { name: kernel.event_listener, event: kernel.controller, method: provideTemplateTwigObjects, priority: -20 }
-            #- { name: kernel.event_listener, event: kernel.terminate, method: startJob }
-            - { name: kernel.event_listener, event: kernel.exception, method: onKernelException }
-            - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest, priority: 110 }
-            #kernel.response or kernel.terminate
     app.twig.extension.stringloader:
         class: Twig_Extension_StringLoader
         tags:
